BOT/Cosmic's Utilites/requirements.txt
aiohttp
humanfriendly
discord.py
